+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add support for managing integration services
 - Fix modification of Firmware
+- Remove empty VHD directory on destroy
 
 ## v0.1.1 2026-05-22